perf: Added Dictionary for faster wallet searching
Changed ints to double as wallet transaction sum can be decimal number. Added Dictionary for faster wallet searching and more memory efficiency.

+++ CryptoWalletChecker/Interfaces.cs
@@ -8,8 +8,8 @@ namespace CryptoWalletChecker
 {
     public interface IMethodsServices
     {
-        void WriteToFile(int number, string wallet);
-        int GetTransactionSum(string wallet);
+        void WriteToFile(double number, string wallet);
+        double GetTransactionSum(string wallet);
         bool IsWalletExists(string wallet);
         void SetVisibilityTrueFalse(VerticalStackLayout vsl1, VerticalStackLayout vsl2);
         bool IsSolanaWallet(string wallet);
